// src/enums/route.enum.ts
export enum routeEnum {
  INITIALPAGE = 'Initialpage',
  LOGINPAGE = 'Loginpage',
  SIGNUPDETAILS = 'signupdetails',
  PROFILEPAGE = 'Profilepage',
  DASHBOARDPAGE = 'Dashboardpage',
  ONBOARD = 'onboard',
  MYGOAL = 'mygoal',
  FORGOTCREDENTIALS = 'forgotcredentials',
  UPDATEPASSWORD = 'updatepassword',
  SIGNUPONBOARDFLOW = 'signuponboardflow',
  SHAREINFO = 'shareinfo',
  ONBOARDWELCOME = 'lazyonboardwelcome',
  PRIVACYPOLICY = 'lazyprivacypolicy',
  CREDITOVERVIEW = 'creditoverview',
  PREPAREDASHBOARD = 'preparedashboard',
  VERIFYEMAIL = 'verifyemail',
  REGISTEREMAIL = 'registeremail',
  REGITSERPASSWORD = 'registerpassword',
  VALIDATIONERROR = 'validationerror',
  NEWPASSWORD = 'newpassword',
  DASHBOARDFORM = 'dashboardform',
  DASHBOARDTAB = 'dashboardtab',
  PARTNERLISTPAGE = 'partnerlist',
  PARTNERPROFILE = 'partnerprofile',
  KBA = 'kbs',
  KBAQUESTIONS = 'kbaQuestions',
  CHANGEEMAIL = 'changeemail',
  ONBOARDCOMPLETE = 'onboarcomplete',
}
export enum tabEnum {
  DASHBOARDPAGE = 'Dashboard',
  PARTNERHOMEPAGE = 'partnerhome',
}
export enum OnboardRouteEnum {
  INITIALPAGE = 'Initialpage',
  NEWC = 'newc',
  CCOMPLETE = 'ccomplete',
  CHARECTERFORM = 'charecterform',
  CAPACITYFORM = 'capacityform',
  CAPITALFORM = 'capitalform',
}